refactor(server): Split files
