+++ engine/infra/server/middleware/cors/middleware.go
@@ -1,30 +1,15 @@
-package server
+package corsmiddleware
 
 import (
-	"context"
 	"net/http"
 	"strconv"
 
 	"github.com/compozy/compozy/pkg/config"
-	"github.com/compozy/compozy/pkg/logger"
 	"github.com/gin-gonic/gin"
 )
 
-// LoggerMiddleware attaches request-scoped logger metadata.
-func LoggerMiddleware(ctx context.Context) gin.HandlerFunc {
-	return func(c *gin.Context) {
-		mgr := config.ManagerFromContext(ctx)
-		log := logger.FromContext(ctx)
-		reqCtx := c.Request.Context()
-		reqCtx = config.ContextWithManager(reqCtx, mgr)
-		reqCtx = logger.ContextWithLogger(reqCtx, log)
-		c.Request = c.Request.WithContext(reqCtx)
-		c.Next()
-	}
-}
-
-// CORSMiddleware applies basic CORS headers based on configuration.
-func CORSMiddleware(cfg config.CORSConfig) gin.HandlerFunc {
+// Middleware applies basic CORS headers based on configuration.
+func Middleware(cfg config.CORSConfig) gin.HandlerFunc {
 	return func(c *gin.Context) {
 		origin := c.Request.Header.Get("Origin")
 		allowed := false
