+++ engine/infra/server/middleware/project/middleware.go
@@ -1,12 +1,12 @@
-package server
+package prmiddleware
 
 import (
 	"github.com/compozy/compozy/engine/core"
 	"github.com/gin-gonic/gin"
 )
 
-// ProjectContextMiddleware injects the project name into the request context.
-func ProjectContextMiddleware(projectName string) gin.HandlerFunc {
+// Middleware injects the project name into the request context.
+func Middleware(projectName string) gin.HandlerFunc {
 	return func(c *gin.Context) {
 		ctx := core.WithProjectName(c.Request.Context(), projectName)
 		c.Request = c.Request.WithContext(ctx)
